feat(mongoose): automatic declarations generation
