+++ .gitignore
@@ -8,6 +8,7 @@ node_modules/
 npm-debug.log
 
 # compiled
+**/*.d.ts
 **/*.js
 **/*.map
