+++ mongoose/.npmignore
@@ -1,5 +1,6 @@
 # TS Files
 src/**/*.ts
+!src/**/*.d.ts
 index.ts
 
 # Example
